@@ -35,42 +35,42 @@
     ]
   },
   "dependencies": {
-    "babel-polyfill": "^6.26.0",
     "cross-env": "5.1.0",
     "fs-extra": "^5.0.0",
-    "react": "16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "16.10.2",
+    "react-dom": "^16.10.2",
     "react-redux": "^5.0.6",
-    "react-router-dom": "4.2.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0",
+    "react-router-dom": "^5.1.2",
+    "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0",
     "zip-folder": "^1.0.0"
   },
   "devDependencies": {
+    "@babel/core": "7.0.0",
+    "@babel/preset-react": "^7.6.3",
+    "@babel/preset-env": "^7.6.3",
     "babel-cli": "6.26.0",
-    "babel-core": "6.26.0",
+    "babel-core": "6.26.3",
     "babel-eslint": "7.2.3",
-    "babel-loader": "7.1.2",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
+    "babel-loader": "8.0.6",
+    "babel-polyfill": "^6.26.0",
     "copy-webpack-plugin": "5.0.4",
     "css-loader": "0.28.7",
     "enzyme": "3.1.0",
     "enzyme-adapter-react-16": "1.0.1",
     "eslint": "6.6.0",
-    "eslint-config-airbnb": "15.1.0",
     "eslint-loader": "^3.0.2",
     "eslint-plugin-react": "^7.16.0",
-    "file-loader": "1.1.4",
+    "file-loader": "4.2.0",
     "html-webpack-plugin": "3.2.0",
     "husky": "0.15.0-beta.2",
     "jest": "21.2.0",
     "mini-css-extract-plugin": "^0.8.0",
     "node-sass": "4.13.0",
     "open-browser-webpack-plugin": "0.0.5",
-    "react-addons-test-utils": "15.6.2",
-    "react-hot-loader": "3.1.1",
+    "react-hot-loader": "^4.12.15",
     "react-test-renderer": "16.0.0",
+     "@hot-loader/react-dom": "^16.8.6",
     "rimraf": "2.6.2",
     "sass-loader": "8.0.0",
     "style-loader": "0.19.0",
